# docs/openapi.yaml
openapi: 3.0.3
info:
  title: BPOS001 POS Module API
  version: 1.0.0
  description: POS service with JWT tenant scoping, POS client selection, shift sessions, sale orders, payments, and Swagger UI.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /pos/select/{id_or_code}:
    get:
      tags: [POS Runtime]
      summary: Select POS client for current merchant/branch and token user
      parameters: [{ in: path, name: id_or_code, required: true, schema: { type: string } }]
      responses: { '200': { description: POS selected } }
  /posclients/{id_or_code}/shift-details:
    get:
      tags: [POS Runtime]
      summary: Get open shift status and auto-close it if it has exceeded 24 hours
      parameters: [{ in: path, name: id_or_code, required: true, schema: { type: string } }]
      responses: { '200': { description: Shift details } }
...